Longitudinal Multi-Bunch Feedback Processor
This will be a feedback processor and diagnostic capture device for
monitoring and controlling the bunch by bunch phase of the stored
electrons in the Diamond storage ring.

This development will be based on the existing TMBF (Transverse
Multi-Bunch Feedback) system currently implemented on the Libera
platform, but will instead use a MicroTCA based platform.

The basic hardware platform consists of a Innovative Integration
FMC-500M dual 500Ms ADC/DAC FMC mounted in a VadaTech AMC525 AMC
carrier card with a Virtex-7 X690T FPGA.  Controller software will
run on a Xeon processor mounted in the same MTCA crate.