Set PCI Class Code to something more sensible
It turns out that a class code of 0000 causes problems with BAR IO
assignment when the FPGA is loaded, presumably this is something to do
with the description "Device was built before Class Code definitions were
finalized".  Have now selected class code 1180, described as "Data
Acquisition and Signal Processing Controllers", "Other".