Ensure that all DMA transfers are on 32-byte boundaries
Unfortunately if we ask for a misaligned transfer the DMA engine simple
locks up.  The 32-byte boundary corresponses to the underly 256-bit DMA
transfer unit, so the alignment constraint is reasonable.

Reading from DDR RAM is not particularly efficient with small block sizes.

We also take care to keep our transfers with the 23 bit limit imposed by
the DMA engine (if the kernel would allow us to allocate a block that
large!)