Provide a bit more FIR gain by adjusting range of bits
Trim two mostly unnecessary bits off the top of the FIR output and widen
the internal FIR path to 25 bits (now a perfect match for the detector
input).

Unfortunately we can now silently overflow in the FIR.  This can be avoided
in software by ensuring that the absolute sum of the FIR is no more than 8,
but really we should try to detect this.