Implement support for variable length detector waveforms
We now only deliver as much data as is actually valid in the waveform.
One tricky detail is that EDM doesn't properly support graphs with variable
length waveforms, so for the moment we have a PV DET:FILL_WAVEFORM_S so
that we can switch the truncation behaviour on and off.  In the end this
will have to go.