Reduce risk of MMS overflow reported at startup
There is a race condition between the first MMS readout and the initial
reset of the MMS error condition.  Strictly speaking this should be fixed
by waiting on a condition during MMS startup to ensure the readout thread
has started, but I don't think the extra complexity of code is needed.