Changes to MMS block placement
By forcing the MMS pblock in a more restrictive way we seem to be guiding
placement in a way which improves timing.

The current MMS pblock is as tight as seems reasonable, and a bit more
timing can be recovered by pulling the left of pblock_x1y2 left by one more
column, but this seems a good starting point.