Add automated conversion of EDM to CSS
This version uses a binary "blob" in the form of a .jar file taken from a
development build of CSS; this should be made into a proper tool at some
point.

At present the binding of TUNE:PREFIX to tune_prefix is not implemented and
the paths to scripts are not correct.