Remove the tune fit alias
This indirection, where we use an external PV to set a macro in the display
manager, really does not work very well.  In particular, porting this to
CSS reveals some glaring problems.

This was a hack that I think we're better off dropping altogether.  Now
instead we use the same $(device):$(axis) prefix for the tune fit results.