Reassign PCIe BAR to AXI mappings
The mapping of address spaces between PCIe and AXI is confusing and it
seems a bit confused.  It looks as if we can safely use lower addresses,
and it looks as if newer versions of the PCIe/AXI bridge can't handle the
high addresses we were using.

So now we have BAR0 => 0x10000 and BAR1 => 0x20000, both of length 0x1000.
Note that using address 0x0 triggers a spurious(?) overlap with the DMA to
PCIe address range.  Still not sure about this one.