Replace std_logic with std_ulogic throughout
The difference between these two VHDL types is that std_ulogic raises an
error if a single line is driven from multiple sources, whereas std_logic
"resolves" this case.  When writing for an FPGA the latter is never what
want, and will eventually lead to a synthesis failure.

It seems that with VHDL 2008 we can use std_ulogic_vector interchangeably
with std_logic_vector without fuss, so it makes sense to remove ALL
references to std_logic and std_logic_vector from the VHDL code.