Change FIR control on overview screen to reduce risk of wrong gain
If the FIR gain control is set to high then the result can be a beam dump.
Unfortunately with an enumeration drop-down this is too easy to do by
mistake.

Therefore the overview panel, main point of contact for operators, is
changed so that the gain can only be changed in 6dB increments.