Don't clean one extra bucket in Hybrid mode
During procedure we saw that the last bunch of every train is badly
affected by the cleaning. The more simple solution is just to add an
extra buncket free of cleaning. That is a dirty solution, but it works
fine and anyway beamlines are not sensitive to an impure bucket just
after a bunch train.