Prevent crash when fetching data with no detectors enabled
As reported here
    https://github.com/DLS-Controls-Private-org/DLS-MBF/issues/5
the IOC would crash if detector data is requested over the socket
interface if no detectors are enabled.

Incidentally, I would expect the same crash when fetching detector data
before any data has been acquired.

The fix is to add an extra stage of error handling.  This has to be done
after the capture lock, so makes things a trifle more complicated.