More work on firmware placement
It looks as if using pblocks to pin down both the axi_pcie3_bridge and its
associated master buffer interface is necessary to help with timing.

I'm hoping that this will finally eliminate the timing error that seems to
be occurring between the PCIE core and its associate BRAM buffers.