Add extra register defs option for hidden group
The extra hierarchy needed to group registers for implementation can be
annoying when it is visible outside the implementation.

This commit adds a new option to support the grouping of
registers while hiding its impact on the rest of the hierarchy.  In
practice the only visible impact of a hidden group is an extra _REGS
entry in the generated VHDL.