Initial conversion from Vivado 2017.4 to 2019.2
The IP upgrade was surprisingly painless.  For some reason the M_DSP_REGS
address width was mistakenly increased from 16 to 17 by the upgrade, but
setting this back again seems to be enough.

Implementation completes with 32ps slack, something we haven't seen for a
long time, and seems to run ok.

Still need to validate scripted complete rebuild.