Handle Tune PLL dropout in tune fitter selector
If the tune fitter is configured to report the Tune PLL tune,
ensure that if Tune PLL drops out (reported by sending NaN) then
the selector is reset (to report the fitted tune).

This ensures that a valid tune can be reported after beam drop-out
and restart before Tune PLL is restarted.